Training/requirements.txt
torch==2.0.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchvision==0.15.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
torchaudio==2.0.0 --extra-index-url https://download.pytorch.org/whl/cu118
neuraloperator
wandb
torch_harmonics
matplotlib
ninja